% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot.R
\name{bar_plot}
\alias{bar_plot}
\title{Bar charts of the concentration of coffee shops in a suburb}
\usage{
bar_plot(suburb)
}
\arguments{
\item{'using}{location as choice of suburb'}
}
\value{
'ggplot2' bar plot showing the number of coffee shops in a suburb
}
\description{
Bar charts of the concentration of coffee shops in a suburb
}
\examples{
bar_plot("Carlton")
}
