% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2.R
\docType{data}
\name{census_suburb}
\alias{census_suburb}
\title{Median age and median income of Melbourne suburbs for the census year 2016 and 2021}
\format{
\subsection{'census_suburb'}{

A data frame with 14 rows and 5 columns.
\describe{
\item{Location}{Name of the suburb}
\item{Median_age_21}{Average age of the suburb in 2021}
\item{Median_income_21}{Average personal income of the suburb in 2021}
\item{Median_age_16}{Average age of the suburb in 2016}
\item{Median_income_16}{Average personal income of the suburb in 2016}
}
}
}
\source{
\url{https://www.abs.gov.au/census/find-census-data/datapacks?release=2021&product=GCP&geography=SAL&header=S}

\url{https://www.abs.gov.au/census/find-census-data/datapacks?release=2016&product=GCP&geography=SSC&header=S}
}
\usage{
census_suburb
}
\description{
Median age and median income of Melbourne suburbs for the census year 2016 and 2021
}
\keyword{datasets}
