% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plot.R
\name{map_plot}
\alias{map_plot}
\title{Map of the concentration of coffee shops in Melbourne suburbs}
\usage{
map_plot(year)
}
\arguments{
\item{'census}{year'}
}
\value{
'ggplot2' map showing the number of coffee shops in Melbourne city
}
\description{
Map of the concentration of coffee shops in Melbourne suburbs
}
\examples{
map_plot(2016)
}
