% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_1.R
\docType{data}
\name{vic_coffee_map}
\alias{vic_coffee_map}
\title{Number of coffee shops in Melbourne from 2002 and 2022}
\format{
\subsection{'vic_coffee_map'}{

A data frame with 279 rows and 5 columns.
\describe{
\item{postcode}{Suburb postcode}
\item{Location}{Name of the suburb}
\item{census year}{Census year (2002 to 2022)}
\item{cafe count}{Records of number of coffee business establishment in Melbourne}
\item{geometry}{Geomtric dimension of the suburb}
}
}
}
\source{
\url{https://data.melbourne.vic.gov.au/explore/dataset/cafes-and-restaurants-with-seating-capacity/export/?location=13,-37.8044,144.9387&basemap=mbs-7a7333}

\url{https://www.arcgis.com/home/item.html?id=153afcf67da145fea4d2f21a69a7ef1d}
}
\usage{
vic_coffee_map
}
\description{
Number of coffee shops in Melbourne from 2002 and 2022
}
\keyword{datasets}
